Move message factory to tests/
